add metric.py in metric module
