+++ dlrover/python/tests/test_metric_monitor.py
@@ -23,13 +23,13 @@ from unittest.mock import patch
 import requests
 
 from dlrover.python.common.constants import GpuMetricEnum, NpuMetricEnum
-from dlrover.python.common.metric import (
+from dlrover.python.common.metric.context import get_job_metric_context
+from dlrover.python.common.metric.metric import (
     GpuMetric,
     GpuNodeMetric,
     NpuMetric,
     NpuNodeMetric,
 )
-from dlrover.python.common.metric.context import get_job_metric_context
 from dlrover.python.common.metric.monitor import (
     GpuMetricMonitor,
     NpuMetricMonitor,
